Trying to distinguish otlp and filelog receivers
but it is not working. afaict the filelog receiver is going through both pipelines.
they receive the same data, all with log.file.path populated. I don't know why the
second pipeline also has the filelog receiver, it does not make sense. I blame helm
and yaml merging, maybe? it's very hard to see the config files for the collectors,
everything is hard